QueryBuilder with extra methods called on it is not reported correctly
I understand that it is hard to evaluate code like this when the things happen in separate methods. The solution may be not to analyze the query builder in this case.

And the message is not very helpful because the displayed portion is similar
```
Strict comparison using === between 'SELECT e FROM…' and 'SELECT e FROM…' will always evaluate to false.
```