Register DBAL types automatically
The goal is to use the extension without having to provide a `objectManagerLoader`.

Because `EntityColumnRule` uses `DescriptorRegistry` and that uses DBAL's `Type::getTypesMap` we
need to make sure that all types are defined there as well.

To make that easy, I quickly hacked the `ReflectionDescriptor` to automatically register the type.

As a bonus, I also did the `UuidTypeDescriptor`.

Now I can run it without having to instantiate the Symfony container.